+++ ui/package.json
@@ -9,6 +9,8 @@
     "@patternfly/react-core": "^3.77.2",
     "@patternfly/react-icons": "^3.14.18",
     "@patternfly/react-styles": "^3.5.7",
+    "@patternfly/react-table": "^2.24.64",
+    "@testing-library/dom": ">=5",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
@@ -17,8 +19,10 @@
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-router-dom": "^5.1.3",
+    "axios": "^0.19.1",
     "d3": "^5.15.0",
     "node-sass": "^4.13.0",
+    "prop-types": "^15.6.1",
     "react": "^16.12.0",
     "react-axe": "^3.3.0",
     "react-d3-graph": "^2.3.0",
@@ -32,7 +36,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "clean-start" : "react-scripts start --reset-cache",
+    "clean-start": "react-scripts start --reset-cache",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
