+++ ui/public/alice.json
@@ -1,0 +1,3 @@
+{
+  "rows": [["one", "two", "three", "four", "five"], ["a", "b", "b", "d", "e"]]
+}
