+++ ui/public/bob.json
@@ -1,0 +1,3 @@
+{
+  "rows": [["alpha", "beta", "gamma", "delta", "epsilon"], ["1", "2", "3", "4", "5"]]
+}
