+++ ui/public/carol.json
@@ -1,0 +1,3 @@
+{
+  "rows": [["alpha", "bravo", "charlie", "delta", "echo"], ["100", "200", "300", "400", "500"]]
+}
