--- ui/public/test.json
@@ -1,77 +1,0 @@
-{
-  "nodes": [
-    {
-      "id": "product",
-      "name": "Product",
-      "_color": "blue"
-    },
-    {
-      "id": "project",
-      "name": "Project",
-      "_color": "green"
-    },
-    {
-      "id": "env",
-      "name": "Environment",
-      "_color": "green"
-    },
-    {
-      "id": "sg",
-      "name": "Security Zone"
-    },
-    {
-      "id": "sg-rule",
-      "name": "Security Zone Rule"
-    },
-    {
-      "id": "safe",
-      "name": "Safe"
-    },
-    {
-      "id": "domain-rec",
-      "name": "Domain Record"
-    },
-    {
-      "id": "domain-name",
-      "name": "Domain Name"
-    },
-    {
-      "id": "Host",
-      "name": "Host"
-    },
-    {
-      "id": "software",
-      "name": "Software"
-    },
-    {
-      "id": "software-log",
-      "name": "Software Log"
-    },
-    {
-      "id": "env-log",
-      "name": "Environment Log"
-    },
-    {
-      "id": "lb",
-      "name": "Load Balancer"
-    }
-  ],
-  "links": [
-    {
-      "sid": "product",
-      "tid": "project"
-    },
-    {
-      "sid": "product",
-      "tid": "env"
-    },
-    {
-      "sid": "env",
-      "tid": "sg"
-    },
-    {
-      "sid": "sg",
-      "tid": "lb"
-    }
-  ]
-}
