+++ ui/src/app/auth/Login.tsx
@@ -1,13 +1,33 @@
-import React, { MouseEvent, useContext, useState } from "react";
+import React, { MouseEvent, useState } from "react";
 import "@patternfly/react-core/dist/styles/base.css";
 import "assets/fonts.css";
 import brandImg from "assets/logo-text_64.png";
 import { BackgroundImageSrc, ListItem, LoginFooterItem, LoginForm, LoginPage } from "@patternfly/react-core";
 import bg_image from "assets/images/bg_4k.jpg";
-import userContext from "context/user-context";
 import { withRouter } from "react-router-dom";
+import { useDispatch } from "react-redux";
+import UserProfile from "data/userProfile";
+import { ACTIONS } from "./authReducer";
 
 const Login: React.ComponentClass<{}> = withRouter((props) => {
+  const dispatch = useDispatch();
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const isValidPassword = () => {
+    return password !== "";
+  };
+  const isValidUsername = () => {
+    return username !== "";
+  };
+  const loginHandler = (event: MouseEvent) => {
+    event.preventDefault();
+    if (isValidUsername() && isValidPassword()) {
+      let user = new UserProfile();
+      user.token = "TODO - read from back end";
+      dispatch({type: ACTIONS.SET_USER, user: user});
+      props.history.push("/");
+    }
+  };
   const listItem = (
     <>
       <ListItem>
@@ -27,22 +47,6 @@ const Login: React.ComponentClass<{}> = withRouter((props) => {
     [BackgroundImageSrc.sm2x]: bg_image,
     [BackgroundImageSrc.xs]: bg_image,
     [BackgroundImageSrc.xs2x]: bg_image
-  };
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const user = useContext(userContext);
-  const isValidPassword = () => {
-    return password !== "";
-  };
-  const isValidUsername = () => {
-    return username !== "";
-  };
-  const loginHandler = (event: MouseEvent) => {
-    event.preventDefault();
-    if (isValidUsername() && isValidPassword()) {
-      user.authenticated = true;
-      props.history.push("/");
-    }
   };
   const loginForm = (
     <LoginForm
