+++ ui/src/app/console/utils/useDocumentTitle.ts
