+++ ui/src/app/console/utils/utils.ts
