--- ui/src/console/ConsoleLayout/ConsoleToolbar.tsx
@@ -1,21 +1,0 @@
-import { Toolbar, ToolbarGroup, ToolbarItem } from "@patternfly/react-core";
-import * as React from "react";
-
-// TODO Not sure this is needed
-
-const ConsoleToolbar = (
-  <Toolbar>
-    <ToolbarGroup>
-      <ToolbarItem>Item 1</ToolbarItem>
-    </ToolbarGroup>
-    <ToolbarGroup>
-      <ToolbarItem>Item 2</ToolbarItem>
-      <ToolbarItem>Item 3</ToolbarItem>
-    </ToolbarGroup>
-    <ToolbarGroup>
-      <ToolbarItem>Item 4</ToolbarItem>
-    </ToolbarGroup>
-  </Toolbar>);
-
-export default ConsoleToolbar;
-
