--- ui/src/console/ConsoleLayout/consoleToolbar.tsx
@@ -1,16 +1,0 @@
-import { Toolbar, ToolbarGroup, ToolbarItem } from "@patternfly/react-core";
-import * as React from "react";
-
-const ConsoleToolbar =
-
-  <Toolbar>
-    <ToolbarGroup>
-      <ToolbarItem>Item 1</ToolbarItem>
-    </ToolbarGroup>
-    <ToolbarGroup>
-      <ToolbarItem>This is annoying !!!</ToolbarItem>
-    </ToolbarGroup>
-  </Toolbar>;
-
-export default ConsoleToolbar;
-
